app/requirements.txt
sqlalchemy
pydantic
aiosqlite
asyncio
Authlib
cryptography
urllib3
fastapi
passlib
python-dateutil
dotenv
itsdangerous
uvicorn
bcrypt
PyJWT
pytz
python-dateutil
oauthlib
fastapi-cli
anyio
async-timeout
asyncio
aiohappyeyeballs
aiohttp
annotated-types
pydantic[email]
pathlib
argon2-cffi
python-multipart
PyMuPDF
python-docx
langchain
openai
langchain-openai
numpy
sentence-transformers
rank-bm25
pysbd
spacy
langchain-text-splitters